Motor sound update
There have been several problems with the motor sound.
The signal was sent to a regular node instead of a Node3D - because of
that the position of the sound was incorrect. This has been corrected
and while doing that I thought a signal might not be necessary in this
case and the sound is computed right here.